backend/pyproject.toml
[tool.poetry]
name = "weather-app"
version = "0.0.1"
description = "API para el proyecto de Can I Jog Today"
authors = ["manu <manloralm@outlook.com>"]
packages = [
    { include = "app"},
]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.2"
uvicorn = "^0.32.0"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.6.0"

[tool.poetry.group.test.dependencies]
httpx = "^0.27.2"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.7.0"
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "asyncio: mark test as async",
]

[tool.poetry.scripts]
start = "app.main:run"
